backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
google-generativeai>=0.3.2
langchain-community>=0.0.13
langchain-huggingface>=0.0.1
sentence-transformers>=2.2.2
faiss-cpu>=1.9.0
numpy>=1.24.3
requests>=2.31.0
torch>=2.1.0
transformers>=4.35.2
nest-asyncio>=1.5.8
python-dotenv>=1.0.0
redis>=5.0.0
flask-limiter>=3.5.0
cryptography>=41.0.0
pillow>=10.0.0
wikipedia>=1.4.0
beautifulsoup4>=4.12.0
apscheduler>=3.10.0
gunicorn>=21.2.0